Moved viewer / data related API to item package
